backend/pyproject.toml
[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
  | snapshots
)/
'''
[tool.isort]
default_section = "THIRDPARTY"
known_first_party = [
    "account",
    "core",
    "grievance",
    "household",
    "id_management",
    "intervention",
    "payment",
    "program",
    "targeting.apps.TargetingConfig",
    "utils",
    "registration_datahub",
    "mptt",
    "django_extensions",
    "registration_data",
    "cash_assist_datahub",
    "mis_datahub",
    "erp_datahub",
    "sanction_list",
]
known_django = "django"
sections = ["FUTURE","STDLIB","DJANGO","THIRDPARTY","FIRSTPARTY","LOCALFOLDER"]
include_trailing_comma = true
skip = "migrations"

[tool.poetry]
name = "HCT MIS Backend"
version = "0.1.22"
description = "HCT MIS is UNICEF's humanitarian cash transfer platform."
authors = ["Tivix"]

[tool.poetry.dependencies]
python = "3.9.1"
Django = "3.2"
django-storages = { version = "1.8", extras = ["azure"] }
django-countries = "7.2.1"
sorl-thumbnail = "12.5.0"
django-model-utils = "4.0.0"
django-phonenumber-field = "4.0.0"
phonenumbers = "8.11.1"
social-auth-app-django = "3.1.0"
graphene-file-upload = "1.2.2"
django-mptt = "0.11.0"
django-extensions = "2.2.6"
elasticsearch-dsl ="^7.0.0"
django-elasticsearch-dsl = "7.1.4"
django-constance = { version = "2.7.0", extras = ["redis"] }
django-auditlog = { git = "https://github.com/Tivix/django-auditlog.git", branch="master" }
django-admin-extra-urls = "^3.3"
unittest-xml-reporting = "3.0.4"
django-multiselectfield = "0.1.12"
Jinja2 = "2.11.2"
psycopg2 = "2.8.4"
pycountry = "20.7.3"
openpyxl = "3.0.3"
xlrd = "1.2.0"
ipdb = "0.13.4"
sentry-sdk = "0.19.5"
openpyxl_image_loader = "1.0.5"
graphene-django = "2.15.0"
django-filter = "2.4"
django-cors-headers = "3.7.0"
gunicorn = "20.0.4"
gevent = "20.9.0"
PyJWT = "1.7.1"
freezegun = "0.3.14"
django-concurrency = "^2.2"
defusedxml= "0.7.0rc1"
django-smart-admin = ">=0.5"
# ---
faker = "5.0.2"
greenlet = "0.4.17"
jedi = "0.17.2"
parso = "0.7.1"
pillow ="7.0.0"
prompt-toolkit = "3.0.8"
ptyprocess = "0.6.0"
pygments = "2.7.3"
pytz = "2020.4"
social-auth-core = "3.3.3"
urllib3 = "1.26.2"
wasmer = "0.4.1"
celery = {extras = ["redis"], version = "^5.0.5"}
django-celery-results = "^2.0.1"
django-celery-beat = "^2.2.0"
black = "^20.8b1"
watchdog = "^2.0.2"
PyYAML = "^5.4.1"
argh = "^0.26.2"
django-sysinfo = "^2.4.0"
django-adminfilters = "^1.3.5"
single-source = "^0.1.5"
django-compressor = "2.4.1"


[tool.poetry.dev-dependencies]
factory-boy = "2.12.0"
snapshottest = "0.6.0"
parameterized = "0.7.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
